pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["andrejrabinov <andryabinov@gmail.com>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-yaml = "^1.2.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
